Test updateGit function directly without full biegunka machinery
