+++ biegunka.nix
@@ -4,8 +4,8 @@
 , filepath, free, hspec, hspec-expectations
 , hspec-expectations-lens, HStringTemplate, lens, meep, mtl
 , optparse-applicative, process, random, resourcet, safecopy
-, semigroups, silently, stdenv, stm, template-haskell, temporary
-, text, transformers, unix
+, semigroups, stdenv, stm, template-haskell, temporary, text
+, transformers, unix
 }:
 mkDerivation {
   pname = "biegunka";
@@ -23,8 +23,8 @@ mkDerivation {
   testDepends = [
     base containers data-default-class directory directory-layout
     filepath free hspec hspec-expectations hspec-expectations-lens lens
-    optparse-applicative process random semigroups silently temporary
-    text transformers unix
+    optparse-applicative process random semigroups temporary text
+    transformers unix
   ];
   homepage = "http://biegunka.budueba.com/";
   description = "Configuration development";
